Bugfix: Kronecker term ignored in spline forces.
The code ignored the kronecker(ktype, 0) or kronecker(ltype, 0)
terms in the contributing terms to NconjtmpI and NconjtmpJ.
The issue was present both in ::bondorder and ::bondorderLJ and
led to energy conservation issues.
It has been fixed by checking for the atom type before entering
the offending calculations and adding clarifying comments.


Former-commit-id: 937cf0b996e5a31ebaeecbe412f03a4abaa508e0