Add a (contrived) molecular example for USER-QUIP
This example showcases the use of different 'special_bonds' settings for
different pair styles, so quip gets all the bonded neighbours but lj can
exclude them if it needs to.

The results have been checked against a pure quip implementation of the
potential; the expected lammps output is included.

DISCLAIMER: This example mixes parameters for methane and silane and is
NOT intended to be a realistic representation of either system.

Former-commit-id: 4e0a249e2779e20e10d6bf363db80ee1b8d363be