Fix installation of lammps.py with CMake
In CMake 3.10 the PythonInterp module defined PYTHON_EXECUTABLE.  Since CMake
3.12 the Python module defines Python_EXECUTABLE.

Since the rest of the code was using PYTHON_EXECUTABLE and expecting it to be
defined, no matter which version, in newer versions of CMake this would lead to
lammps.py not being installed at all.

This commit changes the LAMMPS CMake files to use the newer variable name and
sets Python_EXECUTABLE in older versions if needed.


Former-commit-id: 0afa9e60f97033b1c9626df0394f02b169228cce