adapt behavior of lammps-shell when passing filename without -in/-i flag
To better handle the case of file association. We now split the path into
a directory and basename string and change the current working directory
to that directory and read the file with the basename.
This simplifies the .desktop file and makes the LAMMPS shell behave as
expected on Windows, too.


Former-commit-id: 1382559dbe353ffb8361753c3d9d69d8c8c48a86